отрисовал тост при наличии интернета и отсутствии
