--- src/utils/toast/toast.js
@@ -1,17 +1,0 @@
-const TIME_TO_SHOW_TOAST = 5000;
-
-const container = document.createElement(`div`);
-container.classList.add(`toast-container`);
-document.body.append(container);
-
-export const toast = (message) => {
-  const item = document.createElement(`div`);
-  item.textContent = message;
-  item.classList.add(`toast-item`);
-
-  container.append(item);
-
-  setTimeout(() => {
-    item.remove();
-  }, TIME_TO_SHOW_TOAST);
-};
